Fix the problem of mixing unsigned and signed numbers
Description:use the same number type

Issues:https://gitee.com/openharmony/ark_js_runtime/issues/I5C30M:wq

+++ ecmascript/js_bigint.cpp
@@ -565,11 +565,11 @@ JSHandle<BigInt> BigInt::Uint64ToBigInt(JSThread *thread, const uint64_t &number
 
 int64_t BigInt::ToInt64()
 {
-    int len = GetLength();
+    uint32_t len = GetLength();
     ASSERT(len < 2); // The maximum length of the BigInt data is less 2
     uint64_t value = 0;
     uint32_t *addr = reinterpret_cast<uint32_t *>(&value);
-    for (int index = len - 1; index >= 0; --index) {
+    for (int32_t index = static_cast<int32_t>(len - 1); index >= 0; --index) {
         *(addr + index) = GetDigit(index);
     }
     if (GetSign()) {
